JNS: added adapter, currency rates updating every second

+++ app/src/main/java/com/jo_no/curencyconversionapp/ui/main/CurrencyRates.kt
@@ -2,5 +2,5 @@ package com.jo_no.curencyconversionapp.ui.main
 
 data class CurrencyRate(
     val currency: String,
-    val rate: Long
+    val rate: Float
 )
